refine telemetry wiring and coverage
