+++ crates/domains/ironclaw_telemetry/src/buffered_recorder.rs
@@ -1,24 +1,15 @@
 //! Non-blocking producer port for the telemetry batch worker.
 
-use std::{
-    collections::BTreeMap,
-    num::TryFromIntError,
-    sync::{
-        Arc, Mutex, MutexGuard,
-        atomic::{AtomicU8, AtomicU64, Ordering},
-    },
-    time::Duration,
-};
+use std::{sync::Arc, time::Duration};
 
 use chrono::{DateTime, Datelike, Timelike, Utc};
-use ironclaw_host_api::{ids::TenantId, resource::ResourceScope};
+use ironclaw_host_api::resource::ResourceScope;
 use ironclaw_telemetry_contracts::{
     observation::{
         CollectorInstanceId, MAX_DURABLE_COUNTER, ScopedTelemetryObservation, TelemetryObservation,
     },
     recorder::{RecordOutcome, TelemetryRecorder},
 };
-use tokio::sync::Notify;
 use tokio::sync::mpsc;
 use tokio_util::sync::CancellationToken;
 use uuid::Uuid;
@@ -37,7 +28,13 @@ pub const DEFAULT_TELEMETRY_SHUTDOWN_TIMEOUT: Duration = Duration::from_secs(5);
 /// but no unbounded producer-side state is allocated for a new key.
 pub(crate) const MAX_COVERAGE_SIDE_KEYS: usize = 8_192;
 const MAX_TELEMETRY_TIMESTAMP_YEAR: i32 = 9_999;
-const FAILURE_CLASS_COUNT: usize = 8;
+
+mod diagnostics;
+mod intake;
+
+pub(crate) use diagnostics::{DiagnosticsState, FailureClassCode, checked_atomic_add};
+pub use diagnostics::{TelemetryDiagnostics, TelemetryWriteFailureClass};
+pub(crate) use intake::{CoverageSideDelta, Intake, TenantHourKey};
 
 /// Configuration for the bounded telemetry collector.
 #[derive(Debug, Clone)]
@@ -126,38 +123,6 @@ impl TelemetryClock for SystemTelemetryClock {
     }
 }
 
-/// Typed class for operational repository failures.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u8)]
-pub enum TelemetryWriteFailureClass {
-    StorageAdmission = 1,
-    StoragePoolAdmission = 2,
-    StorageOperation = 3,
-    CounterOverflow = 4,
-    InvalidRecord = 5,
-    InvalidData = 6,
-    ShutdownTimeout = 7,
-    CollectorIdResolution = 8,
-}
-
-impl TelemetryWriteFailureClass {
-    const fn from_repr(value: u8) -> Option<Self> {
-        match value {
-            1 => Some(Self::StorageAdmission),
-            2 => Some(Self::StoragePoolAdmission),
-            3 => Some(Self::StorageOperation),
-            4 => Some(Self::CounterOverflow),
-            5 => Some(Self::InvalidRecord),
-            6 => Some(Self::InvalidData),
-            7 => Some(Self::ShutdownTimeout),
-            8 => Some(Self::CollectorIdResolution),
-            _ => None,
-        }
-    }
-}
-
-pub(crate) type FailureClassCode = TelemetryWriteFailureClass;
-
 #[derive(Debug, Clone, Copy)]
 pub(crate) enum PreflightError {
     SystemScope,
@@ -175,660 +140,6 @@ impl PreflightError {
             | Self::InvalidTimestamp
             | Self::InvalidWindowStart => FailureClassCode::InvalidRecord,
             Self::CounterOutOfRange => FailureClassCode::CounterOverflow,
-        }
-    }
-}
-
-/// Count-only worker and queue diagnostics.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct TelemetryDiagnostics {
-    accepted_observation_count: u64,
-    queue_full_drop_count: u64,
-    closed_drop_count: u64,
-    invalid_observation_count: u64,
-    write_failed_observation_count: u64,
-    repository_failure_count: u64,
-    partial_batch_failure_count: u64,
-    flushed_batch_count: u64,
-    flushed_observation_count: u64,
-    last_batch_size: u64,
-    last_flush_latency_ms: u64,
-    last_failure_class: Option<TelemetryWriteFailureClass>,
-    failure_class_counts: [u64; FAILURE_CLASS_COUNT],
-    shutdown_timeout_count: u64,
-    shutdown_write_loss_count: u64,
-    shutdown_abandoned_observation_count: u64,
-    coverage_key_overflow_count: u64,
-    collector_id_resolution_failure_count: u64,
-}
-
-impl TelemetryDiagnostics {
-    pub const fn accepted_observation_count(self) -> u64 {
-        self.accepted_observation_count
-    }
-    pub const fn queue_full_drop_count(self) -> u64 {
-        self.queue_full_drop_count
-    }
-    pub const fn closed_drop_count(self) -> u64 {
-        self.closed_drop_count
-    }
-    pub const fn invalid_observation_count(self) -> u64 {
-        self.invalid_observation_count
-    }
-    pub const fn write_failed_observation_count(self) -> u64 {
-        self.write_failed_observation_count
-    }
-    pub const fn repository_failure_count(self) -> u64 {
-        self.repository_failure_count
-    }
-    pub const fn partial_batch_failure_count(self) -> u64 {
-        self.partial_batch_failure_count
-    }
-    pub const fn flushed_batch_count(self) -> u64 {
-        self.flushed_batch_count
-    }
-    pub const fn flushed_observation_count(self) -> u64 {
-        self.flushed_observation_count
-    }
-    pub const fn last_batch_size(self) -> u64 {
-        self.last_batch_size
-    }
-    pub const fn last_flush_latency_ms(self) -> u64 {
-        self.last_flush_latency_ms
-    }
-    pub const fn last_failure_class(self) -> Option<TelemetryWriteFailureClass> {
-        self.last_failure_class
-    }
-    pub const fn failure_class_count(self, class: TelemetryWriteFailureClass) -> u64 {
-        let index = class as usize;
-        if index == 0 || index > FAILURE_CLASS_COUNT {
-            0
-        } else {
-            self.failure_class_counts[index - 1]
-        }
-    }
-    pub const fn shutdown_timeout_count(self) -> u64 {
-        self.shutdown_timeout_count
-    }
-    pub const fn shutdown_write_loss_count(self) -> u64 {
-        self.shutdown_write_loss_count
-    }
-    pub const fn shutdown_abandoned_observation_count(self) -> u64 {
-        self.shutdown_abandoned_observation_count
-    }
-    pub const fn coverage_key_overflow_count(self) -> u64 {
-        self.coverage_key_overflow_count
-    }
-    pub const fn collector_id_resolution_failure_count(self) -> u64 {
-        self.collector_id_resolution_failure_count
-    }
-}
-
-pub(crate) struct DiagnosticsState {
-    accepted_observation_count: AtomicU64,
-    queue_full_drop_count: AtomicU64,
-    closed_drop_count: AtomicU64,
-    invalid_observation_count: AtomicU64,
-    write_failed_observation_count: AtomicU64,
-    repository_failure_count: AtomicU64,
-    partial_batch_failure_count: AtomicU64,
-    flushed_batch_count: AtomicU64,
-    flushed_observation_count: AtomicU64,
-    last_batch_size: AtomicU64,
-    last_flush_latency_ms: AtomicU64,
-    last_failure_class: AtomicU8,
-    failure_class_counts: [AtomicU64; FAILURE_CLASS_COUNT],
-    shutdown_timeout_count: AtomicU64,
-    shutdown_write_loss_count: AtomicU64,
-    shutdown_abandoned_observation_count: AtomicU64,
-    coverage_key_overflow_count: AtomicU64,
-    collector_id_resolution_failure_count: AtomicU64,
-}
-
-impl Default for DiagnosticsState {
-    fn default() -> Self {
-        Self {
-            accepted_observation_count: AtomicU64::new(0),
-            queue_full_drop_count: AtomicU64::new(0),
-            closed_drop_count: AtomicU64::new(0),
-            invalid_observation_count: AtomicU64::new(0),
-            write_failed_observation_count: AtomicU64::new(0),
-            repository_failure_count: AtomicU64::new(0),
-            partial_batch_failure_count: AtomicU64::new(0),
-            flushed_batch_count: AtomicU64::new(0),
-            flushed_observation_count: AtomicU64::new(0),
-            last_batch_size: AtomicU64::new(0),
-            last_flush_latency_ms: AtomicU64::new(0),
-            last_failure_class: AtomicU8::new(0),
-            failure_class_counts: std::array::from_fn(|_| AtomicU64::new(0)),
-            shutdown_timeout_count: AtomicU64::new(0),
-            shutdown_write_loss_count: AtomicU64::new(0),
-            shutdown_abandoned_observation_count: AtomicU64::new(0),
-            coverage_key_overflow_count: AtomicU64::new(0),
-            collector_id_resolution_failure_count: AtomicU64::new(0),
-        }
-    }
-}
-
-impl DiagnosticsState {
-    pub(crate) fn snapshot(&self) -> TelemetryDiagnostics {
-        let last_failure_class =
-            TelemetryWriteFailureClass::from_repr(self.last_failure_class.load(Ordering::Relaxed));
-        TelemetryDiagnostics {
-            accepted_observation_count: self.accepted_observation_count.load(Ordering::Relaxed),
-            queue_full_drop_count: self.queue_full_drop_count.load(Ordering::Relaxed),
-            closed_drop_count: self.closed_drop_count.load(Ordering::Relaxed),
-            invalid_observation_count: self.invalid_observation_count.load(Ordering::Relaxed),
-            write_failed_observation_count: self
-                .write_failed_observation_count
-                .load(Ordering::Relaxed),
-            repository_failure_count: self.repository_failure_count.load(Ordering::Relaxed),
-            partial_batch_failure_count: self.partial_batch_failure_count.load(Ordering::Relaxed),
-            flushed_batch_count: self.flushed_batch_count.load(Ordering::Relaxed),
-            flushed_observation_count: self.flushed_observation_count.load(Ordering::Relaxed),
-            last_batch_size: self.last_batch_size.load(Ordering::Relaxed),
-            last_flush_latency_ms: self.last_flush_latency_ms.load(Ordering::Relaxed),
-            last_failure_class,
-            failure_class_counts: std::array::from_fn(|index| {
-                self.failure_class_counts[index].load(Ordering::Relaxed)
-            }),
-            shutdown_timeout_count: self.shutdown_timeout_count.load(Ordering::Relaxed),
-            shutdown_write_loss_count: self.shutdown_write_loss_count.load(Ordering::Relaxed),
-            shutdown_abandoned_observation_count: self
-                .shutdown_abandoned_observation_count
-                .load(Ordering::Relaxed),
-            coverage_key_overflow_count: self.coverage_key_overflow_count.load(Ordering::Relaxed),
-            collector_id_resolution_failure_count: self
-                .collector_id_resolution_failure_count
-                .load(Ordering::Relaxed),
-        }
-    }
-
-    pub(crate) fn increment_accepted(&self) {
-        self.add_counter(&self.accepted_observation_count, 1);
-    }
-    pub(crate) fn increment_queue_full(&self) {
-        self.add_counter(&self.queue_full_drop_count, 1);
-    }
-    pub(crate) fn increment_closed(&self) {
-        self.add_counter(&self.closed_drop_count, 1);
-    }
-    pub(crate) fn add_invalid(&self, count: usize) {
-        let Ok(count) = u64::try_from(count) else {
-            self.record_counter_overflow();
-            return;
-        };
-        self.add_counter(&self.invalid_observation_count, count);
-    }
-    pub(crate) fn add_write_failed(&self, count: usize) {
-        let Ok(count) = u64::try_from(count) else {
-            self.record_counter_overflow();
-            return;
-        };
-        self.add_counter(&self.write_failed_observation_count, count);
-    }
-    pub(crate) fn record_repository_failure(&self, class: FailureClassCode) {
-        self.add_counter(&self.repository_failure_count, 1);
-        self.record_failure(class);
-    }
-
-    pub(crate) fn record_partial_batch_failure(&self) {
-        self.add_counter(&self.partial_batch_failure_count, 1);
-        self.record_failure(FailureClassCode::StorageOperation);
-    }
-
-    pub(crate) fn record_failure(&self, class: FailureClassCode) {
-        let index = class as usize - 1;
-        if checked_atomic_add(&self.failure_class_counts[index], 1).is_err()
-            && class != FailureClassCode::CounterOverflow
-        {
-            self.record_counter_overflow();
-        }
-        self.last_failure_class
-            .store(class as u8, Ordering::Relaxed);
-    }
-    pub(crate) fn record_flush(&self, batch_size: usize, latency_ms: u64) {
-        self.add_counter(&self.flushed_batch_count, 1);
-        let Ok(batch_size) = u64::try_from(batch_size) else {
-            self.record_counter_overflow();
-            return;
-        };
-        self.add_counter(&self.flushed_observation_count, batch_size);
-        self.last_batch_size.store(batch_size, Ordering::Relaxed);
-        self.last_flush_latency_ms
-            .store(latency_ms, Ordering::Relaxed);
-    }
-
-    pub(crate) fn record_shutdown_timeout(&self, abandoned: u64) {
-        self.add_counter(&self.shutdown_timeout_count, 1);
-        self.add_counter(&self.shutdown_write_loss_count, abandoned);
-        self.add_counter(&self.shutdown_abandoned_observation_count, abandoned);
-        self.add_counter(&self.write_failed_observation_count, abandoned);
-        self.record_failure(FailureClassCode::ShutdownTimeout);
-    }
-
-    pub(crate) fn record_coverage_key_overflow(&self) {
-        self.add_counter(&self.coverage_key_overflow_count, 1);
-    }
-
-    pub(crate) fn record_collector_id_resolution_failure(
-        &self,
-        error: &ironclaw_telemetry_contracts::observation::BoundedIdentifierError,
-    ) {
-        self.add_counter(&self.collector_id_resolution_failure_count, 1);
-        self.record_failure(classify_collector_id_error(error));
-    }
-
-    fn add_counter(&self, counter: &AtomicU64, amount: u64) {
-        if checked_atomic_add(counter, amount).is_err() {
-            self.record_counter_overflow();
-        }
-    }
-
-    pub(crate) fn record_counter_overflow(&self) {
-        let index = FailureClassCode::CounterOverflow as usize - 1;
-        let _ = checked_atomic_add(&self.failure_class_counts[index], 1);
-        self.last_failure_class
-            .store(FailureClassCode::CounterOverflow as u8, Ordering::Relaxed);
-    }
-}
-
-fn checked_atomic_add(counter: &AtomicU64, amount: u64) -> Result<u64, u64> {
-    counter.fetch_update(Ordering::AcqRel, Ordering::Acquire, |current| {
-        current.checked_add(amount)
-    })
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub(crate) struct TenantHourKey {
-    pub(crate) tenant_id: TenantId,
-    pub(crate) window_start: DateTime<Utc>,
-}
-
-#[derive(Debug, Clone, Copy, Default)]
-pub(crate) struct CoverageSideDelta {
-    pub(crate) accepted_pending: u64,
-    pub(crate) queue_full_drop_count: u64,
-    pub(crate) closed_drop_count: u64,
-    pub(crate) invalid_drop_count: u64,
-    pub(crate) first_observed_at: Option<DateTime<Utc>>,
-    pub(crate) last_observed_at: Option<DateTime<Utc>>,
-}
-
-#[derive(Debug, Default)]
-pub(crate) struct CoverageSideAccumulator {
-    entries: BTreeMap<TenantHourKey, CoverageSideDelta>,
-}
-
-impl CoverageSideAccumulator {
-    fn record(
-        &mut self,
-        key: TenantHourKey,
-        occurred_at: DateTime<Utc>,
-        update: impl FnOnce(&mut CoverageSideDelta) -> Result<(), ()>,
-    ) -> Result<bool, ()> {
-        if let Some(delta) = self.entries.get_mut(&key) {
-            update(delta)?;
-            delta.record_timestamp(occurred_at);
-            return Ok(true);
-        }
-        if self.entries.len() >= MAX_COVERAGE_SIDE_KEYS {
-            return Ok(false);
-        }
-        let mut delta = CoverageSideDelta::default();
-        update(&mut delta)?;
-        delta.record_timestamp(occurred_at);
-        self.entries.insert(key, delta);
-        Ok(true)
-    }
-
-    fn take_drop_deltas(&mut self) -> BTreeMap<TenantHourKey, CoverageSideDelta> {
-        let mut drops = BTreeMap::new();
-        self.entries.retain(|key, delta| {
-            let drop_delta = CoverageSideDelta {
-                accepted_pending: 0,
-                queue_full_drop_count: delta.queue_full_drop_count,
-                closed_drop_count: delta.closed_drop_count,
-                invalid_drop_count: delta.invalid_drop_count,
-                first_observed_at: delta.first_observed_at,
-                last_observed_at: delta.last_observed_at,
-            };
-            if drop_delta.queue_full_drop_count != 0
-                || drop_delta.closed_drop_count != 0
-                || drop_delta.invalid_drop_count != 0
-            {
-                drops.insert(key.clone(), drop_delta);
-                delta.queue_full_drop_count = 0;
-                delta.closed_drop_count = 0;
-                delta.invalid_drop_count = 0;
-                delta.first_observed_at = None;
-                delta.last_observed_at = None;
-            }
-            delta.accepted_pending != 0
-                || delta.queue_full_drop_count != 0
-                || delta.closed_drop_count != 0
-                || delta.invalid_drop_count != 0
-        });
-        drops
-    }
-
-    fn restore_drop_delta(
-        &mut self,
-        key: TenantHourKey,
-        delta: CoverageSideDelta,
-    ) -> Result<(), ()> {
-        let Some(existing) = self.entries.get(&key) else {
-            if self.entries.len() >= MAX_COVERAGE_SIDE_KEYS {
-                return Err(());
-            }
-            self.entries.insert(key, delta);
-            return Ok(());
-        };
-        let queue_full_drop_count = existing
-            .queue_full_drop_count
-            .checked_add(delta.queue_full_drop_count)
-            .ok_or(())?;
-        let closed_drop_count = existing
-            .closed_drop_count
-            .checked_add(delta.closed_drop_count)
-            .ok_or(())?;
-        let invalid_drop_count = existing
-            .invalid_drop_count
-            .checked_add(delta.invalid_drop_count)
-            .ok_or(())?;
-        let first_observed_at = match (existing.first_observed_at, delta.first_observed_at) {
-            (Some(existing), Some(delta)) => Some(existing.min(delta)),
-            (Some(existing), None) => Some(existing),
-            (None, Some(delta)) => Some(delta),
-            (None, None) => None,
-        };
-        let last_observed_at = match (existing.last_observed_at, delta.last_observed_at) {
-            (Some(existing), Some(delta)) => Some(existing.max(delta)),
-            (Some(existing), None) => Some(existing),
-            (None, Some(delta)) => Some(delta),
-            (None, None) => None,
-        };
-        let existing = self.entries.get_mut(&key).ok_or(())?;
-        existing.queue_full_drop_count = queue_full_drop_count;
-        existing.closed_drop_count = closed_drop_count;
-        existing.invalid_drop_count = invalid_drop_count;
-        existing.first_observed_at = first_observed_at;
-        existing.last_observed_at = last_observed_at;
-        Ok(())
-    }
-
-    fn account_observations(
-        &mut self,
-        keys: impl IntoIterator<Item = TenantHourKey>,
-    ) -> Result<(), ()> {
-        let mut requested = BTreeMap::<TenantHourKey, u64>::new();
-        for key in keys {
-            let entry = requested.entry(key).or_default();
-            *entry = entry.checked_add(1).ok_or(())?;
-        }
-        for (key, count) in &requested {
-            if let Some(delta) = self.entries.get(key)
-                && delta.accepted_pending < *count
-            {
-                return Err(());
-            }
-        }
-        for (key, count) in requested {
-            if let Some(delta) = self.entries.get_mut(&key) {
-                delta.accepted_pending -= count;
-            }
-        }
-        self.entries.retain(|_, delta| {
-            delta.accepted_pending != 0
-                || delta.queue_full_drop_count != 0
-                || delta.closed_drop_count != 0
-                || delta.invalid_drop_count != 0
-        });
-        Ok(())
-    }
-
-    fn take_unpersisted(&mut self) -> BTreeMap<TenantHourKey, u64> {
-        let mut abandoned = BTreeMap::new();
-        for (key, delta) in &self.entries {
-            if delta.accepted_pending != 0 {
-                abandoned.insert(key.clone(), delta.accepted_pending);
-            }
-        }
-        for delta in self.entries.values_mut() {
-            delta.accepted_pending = 0;
-        }
-        self.entries.retain(|_, delta| {
-            delta.queue_full_drop_count != 0
-                || delta.closed_drop_count != 0
-                || delta.invalid_drop_count != 0
-        });
-        abandoned
-    }
-}
-
-impl CoverageSideDelta {
-    fn record_timestamp(&mut self, occurred_at: DateTime<Utc>) {
-        self.first_observed_at = Some(match self.first_observed_at {
-            Some(first) => first.min(occurred_at),
-            None => occurred_at,
-        });
-        self.last_observed_at = Some(match self.last_observed_at {
-            Some(last) => last.max(occurred_at),
-            None => occurred_at,
-        });
-    }
-}
-
-struct IntakeState {
-    sender: mpsc::Sender<ScopedTelemetryObservation>,
-    closed: bool,
-    coverage: CoverageSideAccumulator,
-}
-
-#[derive(Debug)]
-pub(crate) enum IntakeAccountingError {
-    KeyCountOverflow {
-        count: usize,
-        source: TryFromIntError,
-    },
-    PendingCountUnderflow {
-        pending: u64,
-        requested: u64,
-    },
-    CoveragePendingUnderflow,
-}
-
-impl IntakeAccountingError {
-    pub(crate) const fn failure_class(&self) -> FailureClassCode {
-        match self {
-            Self::KeyCountOverflow { count, source } => {
-                debug_assert!(*count > u64::MAX as usize);
-                let _ = source;
-                FailureClassCode::CounterOverflow
-            }
-            Self::PendingCountUnderflow { pending, requested } => {
-                debug_assert!(*pending < *requested);
-                FailureClassCode::CounterOverflow
-            }
-            Self::CoveragePendingUnderflow => FailureClassCode::CounterOverflow,
-        }
-    }
-}
-
-pub(crate) struct Intake {
-    state: Mutex<IntakeState>,
-    notify: Notify,
-    pending_observation_count: AtomicU64,
-}
-
-fn record_coverage_event(
-    state: &mut IntakeState,
-    key: TenantHourKey,
-    occurred_at: DateTime<Utc>,
-    diagnostics: &DiagnosticsState,
-    update: impl FnOnce(&mut CoverageSideDelta) -> Result<(), ()>,
-) {
-    match state.coverage.record(key, occurred_at, update) {
-        Ok(true) => {}
-        Ok(false) => diagnostics.record_coverage_key_overflow(),
-        Err(()) => diagnostics.record_counter_overflow(),
-    }
-}
-
-impl Intake {
-    pub(crate) fn new(sender: mpsc::Sender<ScopedTelemetryObservation>) -> Self {
-        Self {
-            state: Mutex::new(IntakeState {
-                sender,
-                closed: false,
-                coverage: CoverageSideAccumulator::default(),
-            }),
-            notify: Notify::new(),
-            pending_observation_count: AtomicU64::new(0),
-        }
-    }
-
-    fn lock(&self) -> MutexGuard<'_, IntakeState> {
-        match self.state.lock() {
-            Ok(guard) => guard,
-            Err(poisoned) => poisoned.into_inner(),
-        }
-    }
-
-    pub(crate) fn close(&self) {
-        let mut state = self.lock();
-        state.closed = true;
-        drop(state);
-        self.notify.notify_one();
-    }
-
-    pub(crate) fn notified(&self) -> impl std::future::Future<Output = ()> + '_ {
-        self.notify.notified()
-    }
-
-    pub(crate) fn take_drop_deltas(&self) -> BTreeMap<TenantHourKey, CoverageSideDelta> {
-        self.lock().coverage.take_drop_deltas()
-    }
-
-    pub(crate) fn restore_drop_deltas(
-        &self,
-        deltas: impl IntoIterator<Item = (TenantHourKey, CoverageSideDelta)>,
-        diagnostics: &DiagnosticsState,
-    ) {
-        let mut state = self.lock();
-        let mut restored_any = false;
-        for (key, delta) in deltas {
-            if state.coverage.restore_drop_delta(key, delta).is_err() {
-                diagnostics.record_coverage_key_overflow();
-            } else {
-                restored_any = true;
-            }
-        }
-        drop(state);
-        if restored_any {
-            self.notify.notify_one();
-        }
-    }
-
-    pub(crate) fn account_observations(
-        &self,
-        keys: impl IntoIterator<Item = TenantHourKey>,
-    ) -> Result<(), IntakeAccountingError> {
-        let keys: Vec<_> = keys.into_iter().collect();
-        let count = u64::try_from(keys.len()).map_err(|source| {
-            IntakeAccountingError::KeyCountOverflow {
-                count: keys.len(),
-                source,
-            }
-        })?;
-        let mut state = self.lock();
-        if state.coverage.account_observations(keys.clone()).is_err() {
-            return Err(IntakeAccountingError::CoveragePendingUnderflow);
-        }
-        match self.pending_observation_count.fetch_update(
-            Ordering::AcqRel,
-            Ordering::Acquire,
-            |pending| pending.checked_sub(count),
-        ) {
-            Ok(_) => Ok(()),
-            Err(pending) => Err(IntakeAccountingError::PendingCountUnderflow {
-                pending,
-                requested: count,
-            }),
-        }
-    }
-
-    pub(crate) fn take_unpersisted(&self) -> (BTreeMap<TenantHourKey, u64>, u64) {
-        let mut state = self.lock();
-        let unpersisted = state.coverage.take_unpersisted();
-        let pending = self.pending_observation_count.swap(0, Ordering::AcqRel);
-        (unpersisted, pending)
-    }
-
-    pub(crate) fn try_record(
-        &self,
-        observation: ScopedTelemetryObservation,
-        key: TenantHourKey,
-        diagnostics: &DiagnosticsState,
-        preflight: Result<(), PreflightError>,
-    ) -> RecordOutcome {
-        let occurred_at = observation.occurred_at();
-        let mut state = self.lock();
-        if state.closed {
-            record_coverage_event(&mut state, key, occurred_at, diagnostics, |delta| {
-                delta.closed_drop_count = delta.closed_drop_count.checked_add(1).ok_or(())?;
-                Ok(())
-            });
-            diagnostics.increment_closed();
-            drop(state);
-            self.notify.notify_one();
-            return RecordOutcome::DroppedClosed;
-        }
-        if let Err(error) = preflight {
-            record_coverage_event(&mut state, key, occurred_at, diagnostics, |delta| {
-                delta.invalid_drop_count = delta.invalid_drop_count.checked_add(1).ok_or(())?;
-                Ok(())
-            });
-            diagnostics.add_invalid(1);
-            diagnostics.record_failure(error.failure_class());
-            drop(state);
-            self.notify.notify_one();
-            return RecordOutcome::DroppedInvalid;
-        }
-        let outcome = state.sender.try_send(observation);
-        match outcome {
-            Ok(()) => {
-                record_coverage_event(&mut state, key, occurred_at, diagnostics, |delta| {
-                    delta.accepted_pending = delta.accepted_pending.checked_add(1).ok_or(())?;
-                    Ok(())
-                });
-                if checked_atomic_add(&self.pending_observation_count, 1).is_err() {
-                    diagnostics.record_counter_overflow();
-                }
-                diagnostics.increment_accepted();
-                RecordOutcome::Accepted
-            }
-            Err(mpsc::error::TrySendError::Full(_)) => {
-                record_coverage_event(&mut state, key, occurred_at, diagnostics, |delta| {
-                    delta.queue_full_drop_count =
-                        delta.queue_full_drop_count.checked_add(1).ok_or(())?;
-                    Ok(())
-                });
-                diagnostics.increment_queue_full();
-                drop(state);
-                self.notify.notify_one();
-                RecordOutcome::DroppedQueueFull
-            }
-            Err(mpsc::error::TrySendError::Closed(_)) => {
-                record_coverage_event(&mut state, key, occurred_at, diagnostics, |delta| {
-                    delta.closed_drop_count = delta.closed_drop_count.checked_add(1).ok_or(())?;
-                    Ok(())
-                });
-                diagnostics.increment_closed();
-                drop(state);
-                self.notify.notify_one();
-                RecordOutcome::DroppedClosed
-            }
         }
     }
 }
@@ -1075,18 +386,6 @@ pub(crate) fn classify_record_error(error: &crate::RecordError) -> FailureClassC
     }
 }
 
-fn classify_collector_id_error(
-    error: &ironclaw_telemetry_contracts::observation::BoundedIdentifierError,
-) -> FailureClassCode {
-    match error {
-        ironclaw_telemetry_contracts::observation::BoundedIdentifierError::Empty { .. }
-        | ironclaw_telemetry_contracts::observation::BoundedIdentifierError::TooLong { .. }
-        | ironclaw_telemetry_contracts::observation::BoundedIdentifierError::ControlCharacters {
-            ..
-        } => FailureClassCode::CollectorIdResolution,
-    }
-}
-
 pub(crate) fn classify_repository_error(
     error: &crate::TelemetryRepositoryError,
 ) -> FailureClassCode {
@@ -1123,6 +422,8 @@ pub(crate) fn classify_repository_error(
 
 #[cfg(test)]
 mod tests {
+    use std::sync::atomic::{AtomicU64, Ordering};
+
     use super::*;
 
     #[test]
